Replaced usages of 4096 constant with getPageSize

+++ src/IO/MMapReadBufferFromFileDescriptor.cpp
@@ -39,8 +39,8 @@ void MMapReadBufferFromFileDescriptor::init(int fd_, size_t offset, size_t lengt
                 ErrorCodes::CANNOT_ALLOCATE_MEMORY);
 
         BufferBase::set(static_cast<char *>(buf), length, 0);
-        
-        int page_size = ::getPageSize();
+
+        size_t page_size = static_cast<size_t>(::getPageSize());
         ReadBuffer::padded = (length % page_size) > 0 && (length % page_size) <= (page_size - 15);
     }
 }
